python/requirements.txt
Appium-Python-Client==0.22
py==1.4.31
pytest==3.0.1
selenium==2.53.6
